requirements.txt
discord.py
bardapi
